requirements.txt
PyJWT~=2.9.0
requests~=2.32.3
PyGithub~=2.3.0
bs4~=0.0.2
beautifulsoup4~=4.12.3